+++ demisto_sdk/commands/common/errors.py
@@ -9,7 +9,8 @@ from demisto_sdk.commands.common.constants import (BETA_INTEGRATION_DISCLAIMER,
 FOUND_FILES_AND_ERRORS: list = []
 FOUND_FILES_AND_IGNORED_ERRORS: list = []
 
-ALLOWED_IGNORE_ERRORS = ['BA101', 'BA106', 'RP102', 'RP104', 'SC100', 'IF106', 'PA113', 'PA116', 'IN126']
+ALLOWED_IGNORE_ERRORS = ['BA101', 'BA106', 'RP102', 'RP104', 'SC100', 'IF106', 'PA113', 'PA116', 'IN126', 'PB105',
+                         'PB106']
 
 PRESET_ERROR_TO_IGNORE = {
     'community': ['BC', 'CJ', 'DS', 'PA117', 'IN125', 'IN126'],
@@ -109,6 +110,8 @@ ERROR_CODE = {
     "playbook_unhandled_condition": "PB102",
     "playbook_unconnected_tasks": "PB103",
     "invalid_deprecated_playbook": "PB104",
+    "playbook_cant_have_deletecontext_all": "PB105",
+    "using_instance_in_playbook": "PB106",
     "description_missing_in_beta_integration": "DS100",
     "no_beta_disclaimer_in_description": "DS101",
     "no_beta_disclaimer_in_yml": "DS102",
@@ -716,6 +719,11 @@ class Errors:
 
     @staticmethod
     @error_code_decorator
+    def using_instance_in_playbook():
+        return "Playbook should not use specific instance."
+
+    @staticmethod
+    @error_code_decorator
     def playbook_unreachable_condition(task_id, next_task_branch):
         return f'Playbook conditional task with id:{task_id} has task with unreachable ' \
                f'next task condition "{next_task_branch}". Please remove this task or add ' \
@@ -731,6 +739,11 @@ class Errors:
     @error_code_decorator
     def playbook_unconnected_tasks(orphan_tasks):
         return f'The following tasks ids have no previous tasks: {orphan_tasks}'
+
+    @staticmethod
+    @error_code_decorator
+    def playbook_cant_have_deletecontext_all():
+        return 'Playbook can not have DeleteContext script with arg all set to yes.'
 
     @staticmethod
     @error_code_decorator
@@ -998,9 +1011,8 @@ class Errors:
     @staticmethod
     @error_code_decorator
     def invalid_package_structure(invalid_files):
-        return 'You should update the following files to the package format, for further details please visit ' \
-               'https://xsoar.pan.dev/docs/integrations/package-dir. ' \
-               'The files are:\n{}'.format('\n'.join(list(invalid_files)))
+        return 'You should update the following file to the package format, for further details please visit ' \
+               'https://xsoar.pan.dev/docs/integrations/package-dir.'
 
     @staticmethod
     @error_code_decorator