+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+* Added a new validation that checks in playbooks for the usage of `DeleteContext` in **validate** command.
+* Added a new validation that checks in playbooks for the usage of specific instance in **validate** command.
+* Added the **--staged** flag to **validate** command to run on staged files only.
+
+
+# 1.2.7
 * Changed input parameters in **find-dependencies** command.
    - Use ***-i, --input*** instead of ***-p, --path***.
    - Use ***-idp, --id-set-path*** instead of ***-i, --id-set-path***.
@@ -6,7 +12,7 @@
 * Fixed an issue in the **format** command where unnecessary files were not skipped.
 * Fixed an issue in the **update-release-notes** command where the *text* argument was not respected in all cases.
 * Fixed an issue in the **validate** command where a warning about detailed description was given for unified or deprecated integrations.
-* Added the **--staged** flag to **validate** command to run on staged files only.
+* Improved the error returned by the **validate** command when running on files using the old format.
 
 # 1.2.6
 * No longer require setting `DEMISTO_README_VALIDATION` env var to enable README mdx validation. Validation will now run automatically if all necessary node modules are available.